Alloc: split into modules, and handle destructors better
