+++ compiler/hash-alloc/src/collections/mod.rs
@@ -1,0 +1,2 @@
+pub mod row;
+pub mod string;
